fix(cloud): address cloud.aws final-review follow-ups
- awsError.details: populate with the smithy fault (client/server/unknown) for
  API errors instead of always-null (matched the advertised { ...details } shape).
- toPlain: use a provider-neutral wrapped error, not mapGoogleError — it is shared
  by all cloud providers; a (de)serialise failure is a serialisation bug, not a
  google API error.
- parseAWSConfig: reject a credentials object that sets only some of
  accessKeyId/secretAccessKey/sessionToken (static creds need key+secret; a lone
  sessionToken was silently dropped). Add the "runs on event loop" doc comment.
- docs: document that the 3 CloudWatch metric methods are pass-through and take an
  AWS-SDK-shaped PascalCase object; clarify sync-throw (cloud.aws) vs async-reject
  (service methods) in the aws entry.
- cloudwatchlogs: startTime/endTime carrier typed int64 (was int).
- tests: mapAWSError nil/transport/fault-details cases; incomplete-credentials
  rejection; SQS ReceiveMessage mock returns a real MD5OfBody so the receive-side
  checksum path is exercised.

+++ cmd/sercon/cloud_aws_cloudwatchlogs.go
@@ -18,7 +18,7 @@ import (
 type awsCloudWatchLogsArgs struct {
 	prefix, logGroupName, logStreamName, filterPattern, queryString, queryId string
 	limit                                                                    int
-	startTime, endTime                                                       int
+	startTime, endTime                                                       int64
 }
 
 // newCloudWatchLogsClient builds a cloudwatchlogs.Client for cfg. Unlike its
@@ -119,8 +119,8 @@ func awsCloudWatchLogsStartQuery(ctx context.Context, cfg awsConfig, a awsCloudW
 	out, err := c.StartQuery(ctx, &cloudwatchlogs.StartQueryInput{
 		LogGroupName: aws.String(a.logGroupName),
 		QueryString:  aws.String(a.queryString),
-		StartTime:    aws.Int64(int64(a.startTime)),
-		EndTime:      aws.Int64(int64(a.endTime)),
+		StartTime:    aws.Int64(a.startTime),
+		EndTime:      aws.Int64(a.endTime),
 	})
 	if err != nil {
 		return nil, mapAWSError(err)
@@ -158,8 +158,8 @@ func awsCloudWatchLogsExtract(call goja.FunctionCall) (awsCloudWatchLogsArgs, er
 	a.queryString = optString(o, "queryString", "")
 	a.queryId = optString(o, "queryId", "")
 	a.limit = optInt(o, "limit", 0)
-	a.startTime = optInt(o, "startTime", 0)
-	a.endTime = optInt(o, "endTime", 0)
+	a.startTime = int64(optInt(o, "startTime", 0))
+	a.endTime = int64(optInt(o, "endTime", 0))
 	return a, nil
 }
